+++ core-ng/src/main/java/core/framework/impl/web/service/WebServiceControllerBuilder.java
@@ -22,13 +22,13 @@ import static core.framework.impl.code.CodeBuilder.typeVariableLiteral;
 /**
  * @author neo
  */
-public class ServiceControllerBuilder<T> {
+public class WebServiceControllerBuilder<T> {
     private final Class<T> serviceInterface;
     private final T service;
     private final Method method;
     private final HTTPStatus responseStatus;
 
-    public ServiceControllerBuilder(Class<T> serviceInterface, T service, Method method) {
+    public WebServiceControllerBuilder(Class<T> serviceInterface, T service, Method method) {
         this.serviceInterface = serviceInterface;
         this.service = service;
         this.method = method;