+++ core-ng/src/test/java/core/framework/impl/web/service/WebServiceImplValidatorTest.java
@@ -1,0 +1,13 @@
+package core.framework.impl.web.service;
+
+import org.junit.Test;
+
+/**
+ * @author neo
+ */
+public class WebServiceImplValidatorTest {
+    @Test
+    public void validate() {
+        new WebServiceImplValidator<>(TestWebService.class, new TestWebServiceImpl()).validate();
+    }
+}